// apps/portal/src/api/generated/model/voteOptionResponseDto.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * HOA Manager API
 * The HOA Manager API description
 * OpenAPI spec version: 1.0
 */
import type { VoteOptionResponseDtoOptionKey } from "./voteOptionResponseDtoOptionKey";

export interface VoteOptionResponseDto {
    id: string;
    label: string;
    sortOrder: number;
    optionKey: VoteOptionResponseDtoOptionKey;
}